# backend/services/api-gateway/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "PatientRecords API (stub)",
    "version": "0.1.0",
    "description": "Minimal API spec for development â€” includes auth and patient endpoints."
  },
  "servers": [
    { "url": "http://localhost:5000", "description": "Local dev" }
  ],
...
  "paths": {
...
    "/api/patients": {
      "get": {
        "tags": ["Patient Demographics"],
        "summary": "List patients",
        "parameters": [
          { "name": "q", "in": "query", "schema": { "type": "string" }, "description": "Search query" }
        ],
        "responses": { "200": { "description": "A list of patients" }, "401": { "description": "Unauthorized" } },
        "security": [ { "bearerAuth": [] } ]
      }
    },
    "/api/patients/{id}": {
      "get": {
        "tags": ["Patient Demographics"],
        "summary": "Get patient details",
        "parameters": [ { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } } ],
        "responses": { "200": { "description": "Patient object" }, "404": { "description": "Not found" } },
        "security": [ { "bearerAuth": [] } ]
      }
    },
    "/api/patients/{id}/vitals": {
      "get": {
        "tags": ["Vitals"],
        "summary": "Get vitals for a patient",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } },
          { "name": "from", "in": "query", "schema": { "type": "string" }, "description": "Start date (inclusive)" },
          { "name": "to", "in": "query", "schema": { "type": "string" }, "description": "End date (inclusive)" },
          { "name": "type", "in": "query", "schema": { "type": "string" }, "description": "Filter by vital type" }
        ],
        "responses": { "200": { "description": "Array of vitals" }, "404": { "description": "Not found" } },
        "security": [ { "bearerAuth": [] } ]
      },
      "post": {
        "tags": ["Vitals"],
        "summary": "Create new vital record",
        "parameters": [ { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } } ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "dateofobservation": { "type": "string" },
                  "observationcode": { "type": "string" },
                  "observationcodesystem": { "type": "string" },
                  "organizationname": { "type": "string" },
                  "vital_description": { "type": "string" },
                  "unit": { "type": "string" },
                  "value": { "type": "string" },
                  "percentile": { "type": "string" }
                },
                "required": ["dateofobservation", "vital_description"]
              }
            }
          }
        },
        "responses": { "201": { "description": "Vital record created" }, "400": { "description": "Bad request" }, "404": { "description": "Patient not found" } },
        "security": [ { "bearerAuth": [] } ]
      }
    },
    "/api/patients/{id}/labs": {
      "get": {
        "tags": ["Labs"],
        "summary": "Get labs for a patient",
        "parameters": [ { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } } ],
        "responses": { "200": { "description": "Array of labs" }, "404": { "description": "Not found" } },
        "security": [ { "bearerAuth": [] } ]
      },
      "post": {
        "tags": ["Labs"],
        "summary": "Create new lab record",
        "parameters": [ { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } } ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "date": { "type": "string" },
                  "test_name": { "type": "string" },
                  "test_code": { "type": "string" },
                  "result": { "type": "string" },
                  "unit": { "type": "string" },
                  "reference_range": { "type": "string" }
                },
                "required": ["date", "test_name"]
              }
            }
          }
        },
        "responses": { "201": { "description": "Lab record created" }, "400": { "description": "Bad request" }, "404": { "description": "Patient not found" } },
        "security": [ { "bearerAuth": [] } ]
      }
    },
    "/api/patients/{id}/medications": {
      "get": {
        "tags": ["Medications"],
        "summary": "Get medications for a patient",
        "parameters": [ { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } } ],
        "responses": { "200": { "description": "Array of medications" }, "404": { "description": "Not found" } },
        "security": [ { "bearerAuth": [] } ]
      },
      "post": {
        "tags": ["Medications"],
        "summary": "Create new medication",
        "parameters": [ { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } } ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "startDate": { "type": "string" },
                  "name": { "type": "string" },
                  "dose": { "type": "string" },
                  "frequency": { "type": "string" },
                  "indication": { "type": "string" },
                  "route": { "type": "string" }
                },
                "required": ["name"]
              }
            }
          }
        },
        "responses": { "201": { "description": "Medication created" }, "400": { "description": "Bad request" }, "404": { "description": "Patient not found" } },
        "security": [ { "bearerAuth": [] } ]
      }
    },
    "/api/patients/{id}/visits": {
      "get": {
        "tags": ["Visits"],
        "summary": "Get visits for a patient",
        "parameters": [ { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } } ],
        "responses": { "200": { "description": "Array of visits" }, "404": { "description": "Not found" } },
        "security": [ { "bearerAuth": [] } ]
      },
      "post": {
        "tags": ["Visits"],
        "summary": "Create new visit record",
        "parameters": [ { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } } ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "date": { "type": "string" },
                  "visitType": { "type": "string", "enum": ["hospital", "clinic", "office"] },
                  "reason": { "type": "string" },
                  "notes": { "type": "string" },
                  "provider_name": { "type": "string" },
                  "facility_name": { "type": "string" },
                  "discharge_status": { "type": "string" }
                },
                "required": ["date", "visitType"]
              }
            }
          }
        },
        "responses": { "201": { "description": "Visit created" }, "400": { "description": "Bad request" }, "404": { "description": "Patient not found" } },
        "security": [ { "bearerAuth": [] } ]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": { "type": "http", "scheme": "bearer", "bearerFormat": "JWT" }
    }
  }
}